+++ src/View/Helper/HtmlHelper.php
@@ -109,7 +109,7 @@ class HtmlHelper extends CakeHtmlHelper
      *
      * If `$url` is not null, creates a link (`<a>` tag) with the appearance
      *  of a button.
-     * @param string $title Button title
+     * @param string|null $title Button title
      * @param string|array|null $url Cake-relative URL or array of URL
      *  parameters or external URL
      * @param array $options Array of options and HTML attributes
@@ -117,7 +117,7 @@ class HtmlHelper extends CakeHtmlHelper
      * @uses link()
      * @uses tag()
      */
-    public function button($title, $url = null, array $options = [])
+    public function button($title = null, $url = null, array $options = [])
     {
         $options = optionsParser($options, ['role' => 'button'])->addButtonClasses();
 
@@ -395,13 +395,13 @@ class HtmlHelper extends CakeHtmlHelper
 
     /**
      * Creates an HTML link
-     * @param string $title The content to be wrapped by <a> tags
+     * @param string|null $title The content to be wrapped by <a> tags
      * @param string|array|null $url Cake-relative URL or array of URL
      *  parameters or external URL
      * @param array $options Array of options and HTML attributes
      * @return string
      */
-    public function link($title, $url = null, array $options = [])
+    public function link($title = null, $url = null, array $options = [])
     {
         $options = optionsParser($options, ['escape' => false, 'title' => $title]);
         $options->add('title', trim(h(strip_tags($options->get('title')))))->tooltip();